reflect changes to video uploads in the client
In xabbuh/panda-client 1.1.0, possibilities were added to encode a
video only with a subset of the available profiles. Additionally, the
path under which the video was to be retrieved is configurable and a
payload can be stored together with the video. These changes are hereby
reflected in the panda:video:upload command.