Make the HTTPlug layer configurable
By default, the HTTP client is still created automatically by the
PHP-HTTP discovery system called in the library. But is is now possible
to provide an HTTP client explicitly to allow advanced use cases (configuring
some caching for API calls, in case the API supports it for instance).
This configuration relies on configuring the client service externally (there
is a HTTPlugBundle for that) and passing the id in the configuration.